feat: organizes components and creates item wrapper
