+++ resources/js/app.js
@@ -43,14 +43,16 @@ window.Vue = require('vue');
 // files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(key).default))
 
 Vue.component('base-modal', require('./components/base/Modal.vue').default);
-Vue.component('reaction-list', require('./components/ReactionListComponent.vue').default);
-Vue.component('comment-list', require('./components/CommentListComponent.vue').default);
-Vue.component('feedback-form', require('./components/FeedbackFormComponent.vue').default);
-Vue.component('service-form', require('./components/ServiceFormComponent.vue').default);
-Vue.component('item-list', require('./components/ItemListComponent.vue').default);
 Vue.component('in-place-loader', require('./components/base/InPlaceLoader.vue').default);
 Vue.component('popup-loader', require('./components/base/PopupLoader.vue').default);
 
+Vue.component('reaction-list', require('./components/items/ReactionList.vue').default);
+Vue.component('comment-list', require('./components/items/CommentList.vue').default);
+Vue.component('feedback-form', require('./components/items/FeedbackForm.vue').default);
+Vue.component('service-form', require('./components/items/ServiceForm.vue').default);
+Vue.component('item-list', require('./components/items/ItemList.vue').default);
+
+Vue.component('item-wrapper', require('./components/wrappers/ItemWrapper.vue').default);
 /**
  * Next, we will create a fresh Vue application instance and attach it to
  * the page. Then, you may begin adding components to this application
