+++ resources/js/components/items/ReactionList.vue
